Move clause() from CaseExpressionTrait
